+++ controllers/api/post-routes.js
@@ -1,3 +1,5 @@
+const mediaControl = require('../imageModal');
+
 const router = require('express').Router();
 const { Post, User } = require('../../models');
 const withAuth = require('../../utils/auth');
@@ -291,64 +293,8 @@ router.get('/edit-post/:id', withAuth, async (req, res) => {
     })
     const post = editPostData.get({ plain: true });
 
-
-
     if (post.media) {
-
-      // id1,id2,id2 => ['id1','id2','id3']
-      const public_id_list = post.media.split(',');
-      const mediaUrl = [];
-
-      for (let i = 0; i < public_id_list.length; i++) {
-        // const info = await cloudinary.v2.api.resource(public_id_list[i]);
-        // console.log("media info: ",info);
-        const mediaFile = public_id_list[i].split('?');
-        const public_id = mediaFile[0];
-        let resource_type;
-        if (mediaFile.length >= 2) {
-          resource_type = mediaFile[1];
-        }
-
-        if (resource_type === 'video') {
-          const video = await cloudinary.video(public_id, {
-            controls: false,
-            transformation:
-              { height: 100, quality: 50, crop: "scale" },
-            fallback_content: "Your browser does not support HTML5 video tags."
-          });
-
-          const media = {
-            url: video,
-            video: true,
-            id: public_id,
-            resource_type
-          };
-          mediaUrl.push(media);
-        } else if (resource_type === 'raw') {
-          const raw = public_id.split('!')[1];
-          const id = public_id.split('!')[0];
-          const media = {
-            url: raw,
-            raw: true,
-            id: id,
-            resource_type
-          }
-          mediaUrl.push(media);
-        } else {
-          const image = await cloudinary.url(public_id, { transformation: { width: 100, crop: "scale" } });
-          const media = {
-            url: image,
-            image: true,
-            id: public_id,
-            resource_type
-          }
-          mediaUrl.push(media);
-        }
-      }
-
-      post.media = mediaUrl;
-      console.log('new link: ', post.media);
-
+      mediaControl.mediaParse(post, 100, 100, 20);            
     }
 
     res.render('edit-post', { post });
