+++ views/dashboard.handlebars
@@ -90,7 +90,7 @@
     </div>
   </div>
 </div>
-
+{{!-- Modal End --}}
 
 {{#if posts.length}}
 <div class="container post-list" >
