+++ views/homepage.handlebars
@@ -1,13 +1,64 @@
+<!-- Modal -->
+<div class="modal fade" id="exampleModal" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
+  <div class="modal-dialog modal-dialog-centered modal-xl">
+    <div class="modal-content">
+      <div class="modal-header">
+        {{!-- <h5 class="modal-title" id="exampleModalLabel">Modal title</h5> --}}
+        <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close" id="close-modal"></button>
+      </div>
+      <div class="modal-body">
+
+        <div id="carouselExampleControls" class="carousel slide" data-bs-ride="carousel">
+          <div class="carousel-inner" id="image-show">
+            {{!-- image show --}}
+            {{!-- <div class="carousel-item active">
+              <img src="..." class="d-block w-100" alt="...">
+            </div>
+            <div class="carousel-item">
+              <img src="..." class="d-block w-100" alt="...">
+            </div>
+            <div class="carousel-item">
+              <img src="..." class="d-block w-100" alt="...">
+            </div> --}}
+            {{!--  --}}
+          </div>
+          <button class="carousel-control-prev" type="button" data-bs-target="#carouselExampleControls" data-bs-slide="prev">
+            <span class="carousel-control-prev-icon" aria-hidden="true"></span>
+            <span class="visually-hidden">Previous</span>
+          </button>
+          <button class="carousel-control-next" type="button" data-bs-target="#carouselExampleControls" data-bs-slide="next">
+            <span class="carousel-control-next-icon" aria-hidden="true"></span>
+            <span class="visually-hidden">Next</span>
+          </button>
+        </div>
+        {{!-- <div class="container text-center" id="image-show"> --}}
+            {{!-- <img src='https://res.cloudinary.com/drmapjksn/image/upload/c_fill,w_600,h_400/unsigned/pig_jseqbr' alt="image" > --}}
+        {{!-- </div>  --}}
+        
+      </div>
+      {{!-- <div class="modal-footer">
+        <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
+        <button type="button" class="btn btn-primary">Save changes</button>
+      </div> --}}
+    </div>
+  </div>
+</div>
+{{!-- Modal End --}}
+
+
 <!--Homepage displays all posts from all users -->
 <h6>Number of times you've visited this homepage: {{countVisit}}</h6>
 
 
 {{#if posts.length}}
 <div class="container">
-<section class="posts col">
+<section class="posts col" id="post-list-show">
   {{#each posts as |post|}}
   {{> post-details}}
   {{/each}}
 </section>
 {{/if}}
 </div>
+
+
+<script src="/js/homeImages.js"></script>
